+++ CHANGELOG.md
@@ -33,6 +33,7 @@ The releases of Vieb aim to follow [semantic versioning](https://semver.org).
 
 - Resolve file paths with query param characters in them and update the url accordingly
 - Background color override not checking for background images before applying
+- Mouse forward and back buttons not working on Windows
 
 ### Security